requirements.txt
alembic==1.7.5
aniso8601==9.0.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
